@@ -21,33 +21,35 @@
   },
   "license": "MIT",
   "dependencies": {
+    "immutability-helper": "^2.0.0",
     "lodash.throttle": "^3.0.4"
   },
   "devDependencies": {
-    "babel-core": "^5.8.25",
-    "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.3.2",
+    "babel-core": "^6.20.0",
+    "babel-eslint": "7.0.0",
+    "babel-loader": "6.2.9",
+    "babel-plugin-inferno": "1.2.0-beta13",
+    "babel-preset-inferno-app": "1.0.1",
+    "case-sensitive-paths-webpack-plugin": "1.1.4",
     "babel-runtime": "^6.3.13",
     "coding-standards": "git+ssh://git@github.com:60frames/coding-standards.git#1.0.5",
     "css-loader": "^0.21.0",
-    "eslint": "^1.7.3",
-    "eslint-plugin-react": "^3.6.3",
+    "eslint": "^3.8.1",
+    "eslint-config-inferno-app": "^0.3.0",
+    "eslint-loader": "1.6.0",
+    "eslint-plugin-flowtype": "^2.21.0",
+    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-inferno": "6.6.0",
+    "eslint-plugin-jsx-a11y": "2.2.3",
     "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.4",
     "grunt": "^0.4.5",
     "grunt-fontello-react": "^3.0.0",
     "grunt-gh-pages": "^0.10.0",
+    "inferno": "1.0.0-beta32",
+    "inferno-component": "1.0.0-beta32",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.6",
-    "webpack": "^1.12.2",
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0"
-  },
-  "peerDependencies": {
-    "react": ">=0.14.0"
-  },
-  "keywords": [
-    "react-component",
-    "html5-video"
-  ]
+    "webpack": "^1.12.2"
+  }
 }
