@@ -1,39 +1,21 @@
-/*eslint-disable */
-/*
- * Generated by the 'fontello-react' Grunt task.
- */
-import React from 'react';
+/* eslint-disable */
 
-var Icon = React.createClass({
+import Inferno from 'inferno';
+import Component from 'inferno-component'
 
-    propTypes: {
-        name: React.PropTypes.oneOf([
-            'play-1',
-			'volume-off',
-			'volume-down',
-			'volume-up',
-			'resize-full',
-			'resize-small',
-			'pause-1'
-        ])
-    },
-
-    /**
-     * Default the icon to the first one just to show something
-     * @return {Object} The default props
-     */
-    getDefaultProps() {
-        return {
+class Icon extends Component {
+    constructor() {
+        super();
+        this.props = {
             name: 'play-1'
         };
-    },
+    }
 
     render() {
         return (
             <span className={'video-icon video-icon--' + this.props.name}></span>
         );
     }
-
-});
+};
 
 export default Icon;
